requirements.txt
crewai
crewai-tools
langchain-community
langchain
langchain-huggingface
langchain-chroma
langchain-groq
sentence-transformers
python-dotenv
pinecone-client
tavily-python
streamlit
datasets
pypdf
langchain-google-genai
pinecone-client
langchain-pinecone
tenacity
